made modification to seperate crossbrowsers and made integration easy
