+++ conf/remote_url_conf.py
@@ -1,5 +1,7 @@
 #The URLs for connecting to BrowserStack and Sauce Labs.
 
 browserstack_url = "http://hub-cloud.browserstack.com/wd/hub"
+browserstack_app_upload_url = "https://api-cloud.browserstack.com/app-automate/upload"
 saucelabs_url = "https://ondemand.eu-central-1.saucelabs.com:443/wd/hub"
+saucelabs_app_upload_url = 'https://api.eu-central-1.saucelabs.com/v1/storage/upload'
 lambdatest_url = "http://{}:{}@hub.lambdatest.com/wd/hub"
