+++ page_objects/Mobile_Base_Page.py
@@ -3,7 +3,7 @@ Page class that all page models can inherit from
 There are useful wrappers for common Selenium operations
 """
 import unittest,os,inspect
-from .driverfactory import DriverFactory
+from .drivers.driverfactory import DriverFactory
 from .core_helpers.selenium_action_objects import Selenium_Action_Objects
 from .core_helpers.logging_objects import Logging_Objects
 from .core_helpers.remote_objects import Remote_Objects
@@ -72,7 +72,7 @@ class Mobile_Base_Page(Borg,unittest.TestCase, Selenium_Action_Objects, Logging_
 
     def register_driver(self,mobile_os_name,mobile_os_version,device_name,app_package,app_activity,remote_flag,device_flag,app_name,app_path,ud_id,org_id,signing_id,no_reset_flag,appium_version,remote_project_name,remote_build_name):
         "Register the mobile driver"
-        self.driver = self.driver_obj.run_mobile(mobile_os_name,mobile_os_version,device_name,app_package,app_activity,remote_flag,device_flag,app_name,app_path,ud_id,org_id,signing_id,no_reset_flag,appium_version,remote_project_name,remote_build_name)
+        self.driver = self.driver_obj.get_mobile_driver(mobile_os_name,mobile_os_version,device_name,app_package,app_activity,remote_flag,device_flag,app_name,app_path,ud_id,org_id,signing_id,no_reset_flag,appium_version,remote_project_name,remote_build_name)
         self.set_screenshot_dir() # Create screenshot directory
         self.set_log_file()
         self.start()
