+++ page_objects/drivers/local_browsers.py
@@ -1,7 +1,6 @@
 """
 Get the webrivers for local browsers.
 """
-from conf import opera_browser_conf
 from selenium.webdriver.chrome.options import Options
 import sys
 from selenium import webdriver
@@ -33,6 +32,7 @@ class LocalBrowsers():
     @staticmethod
     def opera_local():
         """Get webdriver for opera."""
+        from conf import opera_browser_conf
         try:
             opera_browser_location = opera_browser_conf.location
             options = webdriver.ChromeOptions()
@@ -72,3 +72,31 @@ class LocalBrowsers():
         local_driver = webdriver.Chrome(options=options)
 
         return local_driver
+    
+    def app_details(self, desired_capabilities, app_package, app_activity):
+        desired_capabilities['appPackage'] = app_package
+        desired_capabilities['appActivity'] = app_activity
+        return desired_capabilities
+
+    def app_name(self, desired_capabilities, app_path, app_name):
+        import os
+        desired_capabilities['app'] = os.path.join(app_path, app_name)
+        return desired_capabilities
+
+    def ios_capabilities(self, desired_capabilities, app_package, no_reset_flag, ud_id, org_id, signing_id):
+        desired_capabilities['bundleId'] = app_package
+        desired_capabilities['noReset'] = no_reset_flag
+        if ud_id is not None:
+            desired_capabilities['udid'] = ud_id
+            desired_capabilities['xcodeOrgId'] = org_id
+            desired_capabilities['xcodeSigningId'] = signing_id
+        return desired_capabilities
+    
+    def set_mobile_device(self, mobile_os_name, mobile_os_version, device_name):
+        """Setup the mobile device."""
+        desired_capabilities = {}
+        desired_capabilities['platformName'] = mobile_os_name
+        desired_capabilities['platformVersion'] = mobile_os_version
+        desired_capabilities['deviceName'] = device_name
+
+        return desired_capabilities
